Improve implementation of ShoesViewModel and fix FakeShoeRepository
In FakeShoeRepository a method was private, now fixed.

In order to enable loading of predefined data and clearing of shoe list
from the UI, required methods were implemented. This also required
the implementation of an extension function to force an event notification
of the live data, since list objects wrapped in LiveData do not trigger
an update if the list get modified (but the list object remains the same).