FIX: disable/enable menu item "Load predefined data" correctly
After clearing the list of shoes, the menu item for loading predefined
data was not re-enabled. This was fixed by adding a live data property
to the view model, which tracks the status and is updated from within
the view model. An observer in the view can then enable resp. disable
the menu item accordingly.